+++ src/container/questions/components/TitleWithInterviewStart/components/HeaderButton/components/DeleteButton/types.ts
@@ -1,0 +1,3 @@
+export interface IProps {
+  questionId: number;
+}
