+++ src/container/questions/components/TitleWithInterviewStart/components/HeaderButton/components/HeartButton/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useMutation } from '@tanstack/react-query';
 import { useState } from 'react';
