+++ src/container/questions/components/TitleWithInterviewStart/components/HeaderButton/components/UpdateButton/types.ts
@@ -1,0 +1,8 @@
+import { ICategory } from '@/types/categories';
+
+export interface IProps {
+  questionTitle: string;
+  categoryNames: string[];
+  questionId: number;
+  categories: ICategory[];
+}
